Set readtodc parameter 64bit
In order to fix year 2038 limit.

* i386/i386at/rtc.h (readtodc): Make tp parameter uint64_t *.
* i386/i386at/rtc.c (readtodc): Likewise.
* xen/time.c (readtodc): Likewise.
* i386/i386at/model_dep.c (inittodr): Pass uint64_t pointer to readtodc.

+++ i386/i386at/rtc.c
@@ -146,7 +146,7 @@ dectohexdec(int n)
 }
 
 int
-readtodc(u_int *tp)
+readtodc(uint64_t *tp)
 {
 	struct rtc_st rtclk;
 	time_t n;
